moved category content to another module
