+++ pages/dashboard/content_choose.py
@@ -2,16 +2,16 @@ import flet as ft
 
 from database.connect import DataBase
 from database.requests.req_admins import ReqAdmins
-from database.requests.req_categories import ReqCategory
+from pages.dashboard.content.categories import CategoriesContent
+
+#from pages.dashboard.content.categories import set_CategoriesContent
 from pages.dashboard.elements.admin_elements import AdminRow, AdminHeader
-from pages.dashboard.elements.category_elements import el_category_header, CategoryRow, AddCategoryButton  # el_add_category
-from pages.dashboard.elements.row_width import RowWidth_Admin
 from pages.dashboard.head_elements import header
 from pages.dashboard.types import EnumDashContent
 from pages.style.style import *
 
 
-class Dash_Content():
+class Dash_Content:
 
     def __init__(self, page, user_role):
         self.page = page
@@ -25,57 +25,61 @@ class Dash_Content():
         match content_type:
             case EnumDashContent.CATEGORY:
 
-                error_message = ft.SnackBar(
-                    content=ft.Text('Категория с таким названием уже существует'),
-                    bgcolor=inputBgErrorColor
-                )
-
-                self.body_content.clear()
-                self.content_header = header(label_name="Список Категорий", user_role=self.user_role)
-                self.body_content.append(self.content_header)
+                category_content = CategoriesContent(self)
+                category_content.build()
 
 
-                # resul append to body_content
-                req = ReqCategory()
-                max_length_category = req.get_max_length()
-                name_width = max(max_length_category * 8, 100)  # 7 letter size
-                d_width = {"c1": 80, "c2": name_width, "c3": 150, "c4": 90}
-
-                l_controls = []
-
-                #кнопка "Добавить новую категорию"
-                self.body_content.append(
-                    AddCategoryButton(page=self.page,
-                                      d_width=d_width,
-                                      error_message=error_message,
-                                      l_elements=l_controls,    #передается ссылка на список строк, чтобы к нему добавить новую категорию
-                                      )
-                )
-
-
-                self.body_content.append(el_category_header(d_width))  # table header
-                #---rows---
-                for id, c_name, c_order, p_cnt in req.category_products_cnt():
-                    l_controls.append(CategoryRow(page=self.page,
-                                                  d_width=d_width,
-                                                  error_message=error_message,
-                                                  id=id,
-                                                  p_name=c_name,
-                                                  p_product_cnt=str(p_cnt),
-                                                  p_order=c_order,
-                                                  l_elements=l_controls,
-                                                  )
-                                      )
-                self.body_content.append(ft.Column(
-                    controls=l_controls,
-                    spacing=1,
-                    #height=600,     #scroll не будет работать, если изменить размер окна
-                    scroll=ft.ScrollMode.AUTO,
-                    expand=True
-                ))
-                #--rows-----------
-
-                self.body_content.append(error_message)
+                # error_message = ft.SnackBar(
+                #     content=ft.Text('Категория с таким названием уже существует'),
+                #     bgcolor=inputBgErrorColor
+                # )
+                #
+                # self.body_content.clear()
+                # self.content_header = header(label_name="Список Категорий", user_role=self.user_role)
+                # self.body_content.append(self.content_header)
+                #
+                #
+                # # resul append to body_content
+                # req = ReqCategory()
+                # max_length_category = req.get_max_length()
+                # name_width = max(max_length_category * 8, 100)  # 7 letter size
+                # d_width = {"c1": 80, "c2": name_width, "c3": 150, "c4": 90}
+                #
+                # l_controls = []
+                #
+                # #кнопка "Добавить новую категорию"
+                # self.body_content.append(
+                #     AddCategoryButton(page=self.page,
+                #                       d_width=d_width,
+                #                       error_message=error_message,
+                #                       l_elements=l_controls,    #передается ссылка на список строк, чтобы к нему добавить новую категорию
+                #                       )
+                # )
+                #
+                #
+                # self.body_content.append(el_category_header(d_width))  # table header
+                # #---rows---
+                # for id, c_name, c_order, p_cnt in req.category_products_cnt():
+                #     l_controls.append(CategoryRow(page=self.page,
+                #                                   d_width=d_width,
+                #                                   error_message=error_message,
+                #                                   id=id,
+                #                                   p_name=c_name,
+                #                                   p_product_cnt=str(p_cnt),
+                #                                   p_order=c_order,
+                #                                   l_elements=l_controls,
+                #                                   )
+                #                       )
+                # self.body_content.append(ft.Column(
+                #     controls=l_controls,
+                #     spacing=1,
+                #     #height=600,     #scroll не будет работать, если изменить размер окна
+                #     scroll=ft.ScrollMode.AUTO,
+                #     expand=True
+                # ))
+                # #--rows-----------
+                #
+                # self.body_content.append(error_message)
